Move template fixture to conftest.py
